examples/python: opc: allow setting pixels with a list of ints
Don't force me to split up colors into 3-tuples of RGB, that's not cool.
Allow the pixels list to be ints, which are assumed to be a normal
24-bit RGB color code with the high byte ignored.

Also use big-endian halfwords when packing 2-byte lengths in structs,
rather than splitting up the high and low bytes unnecessarily, and drop
the executable bit from opc.py since it's not a standalone script.

Update solid-color.py to account for this.