server: udev rules and minor tweaks
Add a udev rules file to set the fadecandy USB device permissions to
allow non-root access. Mode 664 with the group "input", because that's
a semi-reasonable group that my laptop's user is already a member of.
As usual, install to /etc/udev/rules.d

When discovering USB devices in the server, wait 100ms for the udev
rules to apply the permission changes, otherwise open fails with an
access denied error if the fadecandy is plugged in while the server is
running.

Also build fcserver with -O2 rather than -Os. No need to optimize for
file size, prefer the normal set of optimizations instead.