server: replace libwebsockets and rapidjson submodules with normal files
Submodules are annoying, it's easier to just import the files directly.
I'd prefer to avoid bundled libraries entirely, but these are pretty old
versions (especially libwebsockets) and it'd be a lot of work to port
fcserver to use up-to-date versions of the libraries.

For libwebsockets, just grab the lib directory, which contains the code
actually used by fcserver. For rapidjson, the submodule already only
included the needed code.